Add files via upload
Created a code to automate the task of creating Pareto Chart .
(A Pareto graph for the Quality department is like a spotlight that highlights the most important problems in a process. It shows which issues are causing the majority of quality problems, helping the team to prioritize where they should focus their efforts for improvement. It's all about identifying the vital few issues among the many, so resources can be allocated effectively to make the most significant impact on quality)